Replaced SAML21 platform by SAMR34 platform.
Contribution provided by The Things Industries.
