+++ .vscode/settings.json
@@ -44,7 +44,7 @@
         "MODULATION":"LORA",
 
         // Target board, the following boards are supported:
-        // NAMote72, NucleoL073 (Default), NucleoL152, NucleoL476, SAML21, SKiM880B, SKiM980A, SKiM881AXL, B-L072Z-LRWAN1.
+        // NAMote72, NucleoL073 (Default), NucleoL152, NucleoL476, SAMR34, SKiM880B, SKiM980A, SKiM881AXL, B-L072Z-LRWAN1.
         "BOARD":"NucleoL073",
 
         // MBED Radio shield selection. (Applies only to Nucleo platforms)
