--- Doc/SAML21-platform.md
@@ -1,9 +1,0 @@
-# SAML21 platform support documents
-
-The supported SAML21 platform is composed of 2 elements:
-1. SAML21-XPRO-B [User Guide](http://ww1.microchip.com/downloads/en/DeviceDoc/Atmel-42405-SAML21-Xplained-Pro_User-Guide.pdf)
-2. SX1276 "Wing Board" [Schematic](SX1276-Wing-Board-(03-91016-RA).pdf)
-
-The SAM L21 family data sheet can be found at [SAM L21 Family Data Sheet](http://www.microchip.com/mymicrochip/filehandler.aspx?ddocname=en590687)
-
-Other useful resources can be found at http://www.microchip.com/ATSAML21-XPRO-B
