+++ cmake/gdb-helper.cmake
@@ -91,7 +91,7 @@ function(generate_vscode_launch_openocd TARGET)
     elseif(BOARD STREQUAL SKiM881AXL)
         set(OPENOCD_INTERFACE stlink-v2.cfg)
         set(OPENOCD_TARGET stm32l0.cfg)
-    elseif(BOARD STREQUAL SAML21)
+    elseif(BOARD STREQUAL SAMR34)
         set(OPENOCD_INTERFACE cmsis-dap.cfg)
         set(OPENOCD_TARGET at91samdXX.cfg)
     endif()
