+++ src/apps/LoRaMac/classB/SAMR34/main.c
@@ -1,7 +1,7 @@
 /*!
  * \file      main.c
  *
- * \brief     LoRaMac classC device implementation
+ * \brief     LoRaMac classB device implementation
  *
  * \copyright Revised BSD License, see section \ref LICENSE.
  *
@@ -23,7 +23,7 @@
  * \author    Marten Lootsma(TWTG) on behalf of Microchip/Atmel (c)2017
  */
 
-/*! \file classC/SAML21/main.c */
+/*! \file classB/SAMR34/main.c */
 
 #include <stdio.h>
 #include "utilities.h"
@@ -47,25 +47,43 @@
 #define USE_TTN_NETWORK                             1
 
 /*!
+ * Uncomment to use the deprecated BeaconTiming MAC command
+ */
+//#define USE_BEACON_TIMING
+
+/*!
  * Enables/Disables the setup of a local Multicast Channel Setup.
  */
 #define LOCAL_MULTICAST_SETUP_ENABLED               0
 
 /*!
- * Defines the application data transmission duty cycle. 5s, value in [ms].
+ * Enables/Disables the ping slot frequency hopping.
  */
-#define APP_TX_DUTYCYCLE                            5000
+#define LOCAL_MULTICAST_SETUP_DISABLE_SLOT_HOP      0
 
 /*!
- * Defines a random delay for application data transmission duty cycle. 1s,
+ * Defines the application data transmission duty cycle. 30s, value in [ms].
+ */
+#define APP_TX_DUTYCYCLE                            30000
+
+/*!
+ * Defines a random delay for application data transmission duty cycle. 5s,
  * value in [ms].
  */
-#define APP_TX_DUTYCYCLE_RND                        1000
+#define APP_TX_DUTYCYCLE_RND                        5000
 
 /*!
  * Default datarate
  */
 #define LORAWAN_DEFAULT_DATARATE                    DR_0
+
+/*!
+ * Default ping slots periodicity
+ *
+ * \remark periodicity is equal to 2^LORAWAN_DEFAULT_PING_SLOT_PERIODICITY seconds
+ *         example: 2^3 = 8 seconds. The end-device will open an Rx slot every 8 seconds.
+ */
+#define LORAWAN_DEFAULT_PING_SLOT_PERIODICITY       0
 
 /*!
  * LoRaWAN confirmed messages
@@ -156,6 +174,11 @@ static TimerEvent_t Led1Timer;
 static TimerEvent_t Led2Timer;
 
 /*!
+ * Timer to handle the state of LED beacon indicator
+ */
+static TimerEvent_t LedBeaconTimer;
+
+/*!
  * Indicates if a new packet can be sent
  */
 static bool NextTx = true;
@@ -176,9 +199,14 @@ static enum eDeviceState
     DEVICE_STATE_START,
     DEVICE_STATE_JOIN,
     DEVICE_STATE_SEND,
+    DEVICE_STATE_REQ_DEVICE_TIME,
+    DEVICE_STATE_REQ_PINGSLOT_ACK,
+    DEVICE_STATE_REQ_BEACON_TIMING,
+    DEVICE_STATE_BEACON_ACQUISITION,
+    DEVICE_STATE_SWITCH_CLASS,
     DEVICE_STATE_CYCLE,
     DEVICE_STATE_SLEEP
-}DeviceState;
+}DeviceState, WakeUpState;
 
 /*!
  * LoRaWAN compliance tests support data
@@ -474,7 +502,7 @@ static void OnTxNextPacketTimerEvent( void* context )
         }
         else
         {
-            DeviceState = DEVICE_STATE_SEND;
+            DeviceState = WakeUpState;
             NextTx = true;
         }
     }
@@ -498,6 +526,17 @@ static void OnLed2TimerEvent( void* context )
     TimerStop( &Led2Timer );
     // Switch LED 1 OFF
     GpioWrite( &Led1, 0 );
+}
+
+/*!
+ * \brief Function executed on Beacon timer Timeout event
+ */
+static void OnLedBeaconTimerEvent( void* context )
+{
+    GpioWrite( &Led1, 1 );
+    TimerStart( &Led2Timer );
+
+    TimerStart( &LedBeaconTimer );
 }
 
 /*!
@@ -809,9 +848,45 @@ static void McpsIndication( McpsIndication_t *mcpsIndication )
 
                         mlmeReq.Type = MLME_DEVICE_TIME;
 
-                        LoRaMacStatus_t status = LoRaMacMlmeRequest( &mlmeReq );
-                        printf( "\n###### ===== MLME-Request - MLME_DEVICE_TIME ==== ######\n" );
-                        printf( "STATUS      : %s\n", MacStatusStrings[status] );
+                        LoRaMacMlmeRequest( &mlmeReq );
+                        WakeUpState = DEVICE_STATE_SEND;
+                        DeviceState = DEVICE_STATE_SEND;
+                    }
+                    break;
+                case 9: // Switch end device Class
+                    {
+                        MibRequestConfirm_t mibReq;
+
+                        mibReq.Type = MIB_DEVICE_CLASS;
+                        // CLASS_A = 0, CLASS_B = 1, CLASS_C = 2
+                        mibReq.Param.Class = ( DeviceClass_t )mcpsIndication->Buffer[1];;
+                        LoRaMacMibSetRequestConfirm( &mibReq );
+
+                        DeviceState = DEVICE_STATE_SEND;
+                    }
+                    break;
+                case 10: // Send PingSlotInfoReq
+                    {
+                        MlmeReq_t mlmeReq;
+
+                        mlmeReq.Type = MLME_PING_SLOT_INFO;
+
+                        mlmeReq.Req.PingSlotInfo.PingSlot.Value = mcpsIndication->Buffer[1];
+
+                        LoRaMacMlmeRequest( &mlmeReq );
+                        WakeUpState = DEVICE_STATE_SEND;
+                        DeviceState = DEVICE_STATE_SEND;
+                    }
+                    break;
+                case 11: // Send BeaconTimingReq
+                    {
+                        MlmeReq_t mlmeReq;
+
+                        mlmeReq.Type = MLME_BEACON_TIMING;
+
+                        LoRaMacMlmeRequest( &mlmeReq );
+                        WakeUpState = DEVICE_STATE_SEND;
+                        DeviceState = DEVICE_STATE_SEND;
                     }
                     break;
                 default:
@@ -858,6 +933,8 @@ static void McpsIndication( McpsIndication_t *mcpsIndication )
  */
 static void MlmeConfirm( MlmeConfirm_t *mlmeConfirm )
 {
+    MibRequestConfirm_t mibReq;
+
     printf( "\n###### ===== MLME-Confirm ==== ######\n" );
     printf( "STATUS      : %s\n", EventInfoStatusStrings[mlmeConfirm->Status] );
     if( mlmeConfirm->Status != LORAMAC_EVENT_INFO_STATUS_OK )
@@ -883,7 +960,11 @@ static void MlmeConfirm( MlmeConfirm_t *mlmeConfirm )
                 printf( "DATA RATE   : DR_%d\n", mibGet.Param.ChannelsDatarate );
                 printf( "\n" );
                 // Status is OK, node has joined the network
-                DeviceState = DEVICE_STATE_SEND;
+#if defined( USE_BEACON_TIMING )
+                DeviceState = DEVICE_STATE_REQ_BEACON_TIMING;
+#else
+                DeviceState = DEVICE_STATE_REQ_DEVICE_TIME;
+#endif
             }
             else
             {
@@ -907,6 +988,62 @@ static void MlmeConfirm( MlmeConfirm_t *mlmeConfirm )
             }
             break;
         }
+        case MLME_DEVICE_TIME:
+        {
+            // Setup the WakeUpState to DEVICE_STATE_SEND. This allows the
+            // application to initiate MCPS requests during a beacon acquisition
+            WakeUpState = DEVICE_STATE_SEND;
+            // Switch to the next state immediately
+            DeviceState = DEVICE_STATE_BEACON_ACQUISITION;
+            NextTx = true;
+            break;
+        }
+        case MLME_BEACON_TIMING:
+        {
+            // Setup the WakeUpState to DEVICE_STATE_SEND. This allows the
+            // application to initiate MCPS requests during a beacon acquisition
+            WakeUpState = DEVICE_STATE_SEND;
+            // Switch to the next state immediately
+            DeviceState = DEVICE_STATE_BEACON_ACQUISITION;
+            NextTx = true;
+            break;
+        }
+        case MLME_BEACON_ACQUISITION:
+        {
+            if( mlmeConfirm->Status == LORAMAC_EVENT_INFO_STATUS_OK )
+            {
+                WakeUpState = DEVICE_STATE_REQ_PINGSLOT_ACK;
+            }
+            else
+            {
+#if defined( USE_BEACON_TIMING )
+                WakeUpState = DEVICE_STATE_REQ_BEACON_TIMING;
+#else
+                WakeUpState = DEVICE_STATE_REQ_DEVICE_TIME;
+#endif
+            }
+            break;
+        }
+        case MLME_PING_SLOT_INFO:
+        {
+            if( mlmeConfirm->Status == LORAMAC_EVENT_INFO_STATUS_OK )
+            {
+                mibReq.Type = MIB_DEVICE_CLASS;
+                mibReq.Param.Class = CLASS_B;
+                LoRaMacMibSetRequestConfirm( &mibReq );
+
+                printf( "\n\n###### ===== Switch to Class B done. ==== ######\n\n" );
+
+                WakeUpState = DEVICE_STATE_SEND;
+                DeviceState = WakeUpState;
+                NextTx = true;
+            }
+            else
+            {
+                WakeUpState = DEVICE_STATE_REQ_PINGSLOT_ACK;
+            }
+            break;
+        }
         default:
             break;
     }
@@ -919,6 +1056,8 @@ static void MlmeConfirm( MlmeConfirm_t *mlmeConfirm )
  */
 static void MlmeIndication( MlmeIndication_t *mlmeIndication )
 {
+    MibRequestConfirm_t mibReq;
+
     if( mlmeIndication->Status != LORAMAC_EVENT_INFO_STATUS_BEACON_LOCKED )
     {
         printf( "\n###### ===== MLME-Indication ==== ######\n" );
@@ -932,6 +1071,47 @@ static void MlmeIndication( MlmeIndication_t *mlmeIndication )
         case MLME_SCHEDULE_UPLINK:
         {// The MAC signals that we shall provide an uplink as soon as possible
             OnTxNextPacketTimerEvent( NULL );
+            break;
+        }
+        case MLME_BEACON_LOST:
+        {
+            mibReq.Type = MIB_DEVICE_CLASS;
+            mibReq.Param.Class = CLASS_A;
+            LoRaMacMibSetRequestConfirm( &mibReq );
+
+            printf( "\n\n###### ===== Switch to Class A done. ==== ######\n\n" );
+
+            // Switch to class A again
+#if defined( USE_BEACON_TIMING )
+            WakeUpState = DEVICE_STATE_REQ_BEACON_TIMING;
+#else
+            WakeUpState = DEVICE_STATE_REQ_DEVICE_TIME;
+#endif
+            TimerStop( &LedBeaconTimer );
+            printf( "\n###### ===== BEACON LOST ==== ######\n" );
+            break;
+        }
+        case MLME_BEACON:
+        {
+            if( mlmeIndication->Status == LORAMAC_EVENT_INFO_STATUS_BEACON_LOCKED )
+            {
+                TimerStart( &LedBeaconTimer );
+                printf( "\n###### ===== BEACON %lu ==== ######\n", mlmeIndication->BeaconInfo.Time.Seconds );
+                printf( "GW DESC     : %d\n", mlmeIndication->BeaconInfo.GwSpecific.InfoDesc );
+                printf( "GW INFO     : " );
+                PrintHexBuffer( mlmeIndication->BeaconInfo.GwSpecific.Info, 6 );
+                printf( "\n" );
+                printf( "FREQ        : %lu\n", mlmeIndication->BeaconInfo.Frequency );
+                printf( "DATA RATE   : DR_%d\n", mlmeIndication->BeaconInfo.Datarate );
+                printf( "RX RSSI     : %d\n", mlmeIndication->BeaconInfo.Rssi );
+                printf( "RX SNR      : %d\n", mlmeIndication->BeaconInfo.Snr );
+                printf( "\n" );
+            }
+            else
+            {
+                TimerStop( &LedBeaconTimer );
+                printf( "\n###### ===== BEACON NOT RECEIVED ==== ######\n" );
+            }
             break;
         }
         default:
@@ -996,13 +1176,17 @@ int main( void )
     uint32_t localMcAddress = 0x01020304;
 
     //                               AS923,     AU915,     CN470,     CN779,     EU433,     EU868,     KR920,     IN865,     US915,     RU864
+#if( LOCAL_MULTICAST_SETUP_DISABLE_SLOT_HOP == 1 )
     const uint32_t frequencies[] = { 923200000, 923300000, 505300000, 786000000, 434665000, 869525000, 921900000, 866550000, 923300000, 869100000 };
+#else
+    const uint32_t frequencies[] = { 923200000, 0        , 0,         786000000, 434665000, 869525000, 921900000, 866550000, 0,         69100000 };
+#endif
     const int8_t dataRates[]     = { DR_2,      DR_2,      DR_0,      DR_0,      DR_0,      DR_0,      DR_0,      DR_0,      DR_0,      DR_0 };
 
     McChannelParams_t channel =
     {
         .IsRemotelySetup = false,
-        .Class = CLASS_C,
+        .Class = CLASS_B,
         .IsEnabled = true,
         .GroupID = MULTICAST_0_ADDR,
         .Address = localMcAddress,
@@ -1016,10 +1200,11 @@ int main( void )
         },
         .FCountMin = 0,
         .FCountMax = UINT32_MAX,
-        .RxParams.ClassC =
+        .RxParams.ClassB =
         {
             .Frequency = frequencies[ACTIVE_REGION],
             .Datarate = dataRates[ACTIVE_REGION],
+            .Periodicity = REGION_COMMON_DEFAULT_PING_SLOT_PERIODICITY,
         }
     };
 
@@ -1076,8 +1261,9 @@ int main( void )
 #endif
 
     DeviceState = DEVICE_STATE_RESTORE;
+    WakeUpState = DEVICE_STATE_START;
 
-    printf( "###### ===== ClassC demo application v1.0.0 ==== ######\n\n" );
+    printf( "###### ===== ClassB demo application v1.0.0 ==== ######\n\n" );
 
     while( 1 )
     {
@@ -1150,6 +1336,9 @@ int main( void )
 
                 TimerInit( &Led2Timer, OnLed2TimerEvent );
                 TimerSetValue( &Led2Timer, 25 );
+
+                TimerInit( &LedBeaconTimer, OnLedBeaconTimerEvent );
+                TimerSetValue( &LedBeaconTimer, 5000 );
 
                 mibReq.Type = MIB_PUBLIC_NETWORK;
                 mibReq.Param.EnablePublicNetwork = LORAWAN_PUBLIC_NETWORK;
@@ -1240,25 +1429,87 @@ int main( void )
                 mibReq.Param.NetworkActivation = ACTIVATION_TYPE_ABP;
                 LoRaMacMibSetRequestConfirm( &mibReq );
 
-                DeviceState = DEVICE_STATE_SEND;
+#if defined( USE_BEACON_TIMING )
+                DeviceState = DEVICE_STATE_REQ_BEACON_TIMING;
+#else
+                DeviceState = DEVICE_STATE_REQ_DEVICE_TIME;
+#endif
 #else
                 JoinNetwork( );
 #endif
+                break;
+            }
+            case DEVICE_STATE_REQ_DEVICE_TIME:
+            {
+                MlmeReq_t mlmeReq;
+
+                if( NextTx == true )
+                {
+                    mlmeReq.Type = MLME_DEVICE_TIME;
+
+                    if( LoRaMacMlmeRequest( &mlmeReq ) == LORAMAC_STATUS_OK )
+                    {
+                        WakeUpState = DEVICE_STATE_SEND;
+                    }
+                }
+                DeviceState = DEVICE_STATE_SEND;
+                break;
+            }
+            case DEVICE_STATE_REQ_BEACON_TIMING:
+            {
+                MlmeReq_t mlmeReq;
+
+                if( NextTx == true )
+                {
+                    mlmeReq.Type = MLME_BEACON_TIMING;
+
+                    if( LoRaMacMlmeRequest( &mlmeReq ) == LORAMAC_STATUS_OK )
+                    {
+                        WakeUpState = DEVICE_STATE_SEND;
+                    }
+                }
+                DeviceState = DEVICE_STATE_SEND;
+                break;
+            }
+            case DEVICE_STATE_BEACON_ACQUISITION:
+            {
+                MlmeReq_t mlmeReq;
+
+                if( NextTx == true )
+                {
+                    mlmeReq.Type = MLME_BEACON_ACQUISITION;
+
+                    LoRaMacMlmeRequest( &mlmeReq );
+                    NextTx = false;
+                }
+                DeviceState = DEVICE_STATE_SEND;
+                break;
+            }
+            case DEVICE_STATE_REQ_PINGSLOT_ACK:
+            {
+                MlmeReq_t mlmeReq;
+
+                if( NextTx == true )
+                {
+                    mlmeReq.Type = MLME_LINK_CHECK;
+                    LoRaMacMlmeRequest( &mlmeReq );
+
+                    mlmeReq.Type = MLME_PING_SLOT_INFO;
+                    mlmeReq.Req.PingSlotInfo.PingSlot.Fields.Periodicity = LORAWAN_DEFAULT_PING_SLOT_PERIODICITY;
+                    mlmeReq.Req.PingSlotInfo.PingSlot.Fields.RFU = 0;
+
+                    if( LoRaMacMlmeRequest( &mlmeReq ) == LORAMAC_STATUS_OK )
+                    {
+                        WakeUpState = DEVICE_STATE_SEND;
+                    }
+                }
+                DeviceState = DEVICE_STATE_SEND;
                 break;
             }
             case DEVICE_STATE_SEND:
             {
                 if( NextTx == true )
                 {
-                    mibReq.Type = MIB_DEVICE_CLASS;
-                    LoRaMacMibGetRequestConfirm( &mibReq );
-
-                    if( mibReq.Param.Class!= CLASS_C )
-                    {
-                        mibReq.Param.Class = CLASS_C;
-                        LoRaMacMibSetRequestConfirm( &mibReq );
-                    }
-
                     PrepareTxFrame( AppPort );
 
                     NextTx = SendFrame( );
