+++ src/boards/SAMR34/rtc-board.c
