+++ src/boards/mcu/saml21/hal/src/hal_ext_irq.c
@@ -43,7 +43,7 @@
 
 #include "hal_ext_irq.h"
 
-#define EXT_IRQ_AMOUNT 6
+#define EXT_IRQ_AMOUNT 4
 
 /**
  * \brief Driver version
